+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [#326] Fixed a problem where the CAS integration test step failed in the multinode Jenkins pipeline because there was no OIDC provider.
+  - Setup Keycloak as an OIDC provider in the multinode Jenkins pipeline and use it for the CAS integration tests.
+  - Setup the Test Realm in Keycloak with a test client and test user for the CAS integration tests.
 
 ## [v7.2.7-16] - 2026-05-12
 ### Fixed